Refactor

+++ SpoVM/lab_2/Parent.cpp
@@ -8,50 +8,54 @@
 #include <iostream>
 #include <cstdio>
 #include <list>
+#include <pthread.h>
 
-const int PROC_DELAY = 1;
+void *sendSignals(void *arg);
 
-using namespace std;
+int startThread(std::pair<sigset_t &, std::list<pid_t> &> &args);
+
+void initSigset(sigset_t &set);
 
 int main(int argc, char **argv) {
-    list<pid_t> pidList;
-
-    int receiver = 0;
+    std::list<pid_t> pidList;
     sigset_t waitSet;
-    sigemptyset(&waitSet);
-    sigaddset(&waitSet, SIGUSR2);
-    sigprocmask(SIG_BLOCK, &waitSet, nullptr);
 
-    cout << "Hello! This is the parent process." << endl;
-    cout << "If you want to create new process, please press '+'" << endl;
-    cout << "If you want to delete last process, please press '-'" << endl;
-    cout << "If you want to send signals, please press 'c'" << endl;
-    cout << "If you want to quit, please press 'q'" << endl;
+    initSigset(waitSet);
+
+    std::pair<sigset_t &, std::list<pid_t> &> args(waitSet, pidList);
+    if (startThread(args)) {
+        std::cout << "Error while creation thread" << std::endl;
+        exit(1);
+    }
+
+    std::cout << "Hello! This is the parent process." << std::endl;
+    std::cout << "If you want to create new process, please press '+'" << std::endl;
+    std::cout << "If you want to delete last process, please press '-'" << std::endl;
+    std::cout << "If you want to quit, please press 'q'" << std::endl;
 
     char symbol;
     while (true) {
-        cin.get(symbol);
+        std::cin.get(symbol);
         switch (symbol) {
             case '+':
                 pid_t pid;
                 pid = fork();
                 switch (pid) {
                     case -1:
-                        cout << "Error while creating child process! (fork)" << endl;
+                        std::cout << "Error while creating child process!" << std::endl;
                         exit(EXIT_FAILURE);
                     case 0:
-                        execv("/home/ilyshka/Desktop/Unik/SpoVM/lab_2/cmake-build-debug/lab_2_child", argv);
+                        execv(argv[1], argv);
                         break;
                     default:
                         pidList.push_back(pid);
-                        sleep(PROC_DELAY);
                         break;
                 }
                 break;
 
             case '-':
                 if (pidList.empty()) {
-                    cout << "There are no children to delete!" << endl;
+                    std::cout << "There are no children to delete!" << std::endl;
                 } else {
                     kill(pidList.back(), SIGKILL);
                     pidList.pop_back();
@@ -67,14 +71,35 @@ int main(int argc, char **argv) {
                 }
                 return 0;
 
-            case 'c':
-                for (auto &childPid : pidList) {
-                    kill(childPid, SIGUSR1);
-                    sigwait(&waitSet, &receiver);
-                }
-
             default:
                 continue;
         }
+    }
+}
+
+void initSigset(sigset_t &set) {
+    sigemptyset(&set);
+    sigaddset(&set, SIGUSR2);
+    sigprocmask(SIG_BLOCK, &set, nullptr);
+}
+
+int startThread(std::pair<sigset_t &, std::list<pid_t> &> &args) {
+    pthread_t thread;
+    pthread_attr_t attr;
+    pthread_attr_init(&attr);
+    return pthread_create(&thread, &attr, sendSignals, &args);
+}
+
+void *sendSignals(void *arg) {
+    auto params = static_cast<std::pair<sigset_t &, std::list<pid_t> &> *> (arg);
+    auto &waitSet = params->first;
+    auto &pidList = params->second;
+    while (true) {
+        for (auto &childPid : pidList) {
+            kill(childPid, SIGUSR1);
+            int sig;
+            sigwait(&waitSet, &sig);
+        }
+        sleep(1);
     }
 }
